requirements.txt
paddlepaddle
paddleocr
pywinauto
pillow
pytz
pymongo